requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit-marquee
pillow
streamlit-option-menu
